refactor: change Life component to add buttons for increase/decrease of life and mana, also added 'armor' because removed the Level component
