--- src/components/Level/Level.css
@@ -1,9 +1,0 @@
-.level {
-  width: 30vw;
-  display: flex;
-  justify-content: space-around;
-  border: 2px solid saddlebrown;
-  border-top-right-radius: 16px;
-  border-bottom-left-radius: 16px;
-  margin-top: 10px;
-}
