+++ src/components/Life/Life.css
@@ -1,8 +1,34 @@
-.life {
-  width: 30vw;
+.status {
+  width: 90vw;
   display: flex;
   justify-content: space-around;
   border: 2px solid saddlebrown;
   border-top-right-radius: 16px;
   border-bottom-left-radius: 16px;
+}
+
+.status-info {
+  width: 30vw;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.buttons {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  margin-bottom: 10px;
+}
+
+.buttons > button {
+  background-color: antiquewhite;
+  color: saddlebrown;
+  border: 1px solid saddlebrown;
+  cursor: pointer;
+}
+
+.buttons > button:hover {
+  opacity: 0.8;
 }
